move some of the files to make sense
