+++ lib/bright/sis_apis/one_roster/infinite_campus.rb
@@ -1,5 +1,3 @@
-require_relative "one_roster"
-
 module Bright
   module SisApi
     class OneRoster::InfiniteCampus < OneRoster
